Commit: Interaction with movies is separate; Delete user resolver.
