+++ src/user/user.module.ts
@@ -2,12 +2,11 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { FavoriteGenre, FavoriteMovie, User } from './user.entity';
 import { UserService } from './user.service';
-import { UserResolver } from './user.resolver';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, FavoriteMovie, FavoriteGenre])],
   controllers: [],
-  providers: [UserService, UserResolver],
+  providers: [UserService],
   exports: [UserService],
 })
 export class UserModule {}
